* commit 'e3c9a209dcc8d01973cadf42f2c8c8c77fcfd734':
  Streaming upload

+++ dataset-service/src/main/java/org/eea/dataset/controller/DatasetControllerImpl.java
@@ -11,9 +11,11 @@ import java.util.List;
 import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import org.apache.commons.fileupload.FileItemFactory;
 import org.apache.commons.fileupload.FileItemIterator;
 import org.apache.commons.fileupload.FileItemStream;
 import org.apache.commons.fileupload.FileUploadException;
+import org.apache.commons.fileupload.disk.DiskFileItemFactory;
 import org.apache.commons.fileupload.servlet.ServletFileUpload;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
@@ -283,7 +285,8 @@ public class DatasetControllerImpl implements DatasetController {
   @Override
   @HystrixCommand
   // Put lock
-  @PostMapping("/v1/{datasetId}/streamImportFileData")
+  @PostMapping(path = "/v1/{datasetId}/streamImportFileData",
+      consumes = {MediaType.MULTIPART_FORM_DATA_VALUE})
   @ApiOperation(value = "Import file to dataset data",
       notes = "Allowed roles: \n\n Reporting dataset: LEAD REPORTER, REPORTER WRITE, NATIONAL COORDINATOR \n\n Data collection: CUSTODIAN, STEWARD\n\n Test dataset: CUSTODIAN, STEWARD, STEWARD SUPPORT\n\n Reference dataset: CUSTODIAN, STEWARD\n\n Design dataset: CUSTODIAN, STEWARD, EDITOR WRITE\n\n EU dataset: CUSTODIAN, STEWARD")
   @PreAuthorize("secondLevelAuthorize(#datasetId,'DATASET_LEAD_REPORTER','DATASET_REPORTER_WRITE','DATASCHEMA_STEWARD','DATASCHEMA_CUSTODIAN','DATASCHEMA_EDITOR_WRITE','DATASCHEMA_EDITOR_READ','EUDATASET_CUSTODIAN','DATASET_NATIONAL_COORDINATOR','TESTDATASET_CUSTODIAN','TESTDATASET_STEWARD_SUPPORT','TESTDATASET_STEWARD','REFERENCEDATASET_CUSTODIAN','REFERENCEDATASET_STEWARD') OR checkApiKey(#dataflowId,#providerId,#datasetId,'DATASET_LEAD_REPORTER','DATASET_REPORTER_WRITE','DATASCHEMA_STEWARD','DATASCHEMA_CUSTODIAN','DATASCHEMA_EDITOR_WRITE','EUDATASET_CUSTODIAN','EUDATASET_STEWARD','DATASET_NATIONAL_COORDINATOR','TESTDATASET_CUSTODIAN','TESTDATASET_STEWARD_SUPPORT','TESTDATASET_STEWARD','REFERENCEDATASET_CUSTODIAN','REFERENCEDATASET_STEWARD')")
@@ -308,35 +311,42 @@ public class DatasetControllerImpl implements DatasetController {
           example = ",") @RequestParam(value = "delimiter", required = false) String delimiter,
       @RequestBody HttpServletRequest request) {
 
+    boolean isMultipart = ServletFileUpload.isMultipartContent(request);
 
-    String filename = "";
-    try {
-
+    if (isMultipart) {
+      FileItemFactory factory = new DiskFileItemFactory();
       // Create a new file upload handler
-      ServletFileUpload upload = new ServletFileUpload();
+      ServletFileUpload upload = new ServletFileUpload(factory);
+      String filename = "";
+      try {
 
-      // Parse the request
-      FileItemIterator iter = upload.getItemIterator(request);
-      while (iter.hasNext()) {
-        FileItemStream item = iter.next();
-        try (InputStream stream = item.openStream()) {
-          if (!item.isFormField()) {
-            filename = item.getName();
-            fileTreatmentHelper.importFileDataV2(datasetId, tableSchemaId, stream, replace,
-                integrationId, delimiter, filename);
+
+
+        // Parse the request
+        FileItemIterator iter = upload.getItemIterator(request);
+        while (iter.hasNext()) {
+          FileItemStream item = iter.next();
+          try (InputStream stream = item.openStream()) {
+            if (!item.isFormField()) {
+              filename = item.getName();
+              fileTreatmentHelper.importFileDataV2(datasetId, tableSchemaId, stream, replace,
+                  integrationId, delimiter, filename);
+            }
           }
         }
+      } catch (EEAException | IOException | FileUploadException e) {
+        LOG_ERROR.error(
+            "File import failed: datasetId={}, tableSchemaId={}, fileName={}. Message: {}",
+            datasetId, tableSchemaId, filename, e.getMessage(), e);
+        Map<String, Object> importFileData = new HashMap<>();
+        importFileData.put(LiteralConstants.SIGNATURE, LockSignature.IMPORT_FILE_DATA.getValue());
+        importFileData.put(LiteralConstants.DATASETID, datasetId);
+        lockService.removeLockByCriteria(importFileData);
+        throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
+            EEAErrorMessage.IMPORTING_FILE_DATASET);
       }
-    } catch (EEAException | IOException | FileUploadException e) {
-      LOG_ERROR.error(
-          "File import failed: datasetId={}, tableSchemaId={}, fileName={}. Message: {}", datasetId,
-          tableSchemaId, filename, e.getMessage(), e);
-      Map<String, Object> importFileData = new HashMap<>();
-      importFileData.put(LiteralConstants.SIGNATURE, LockSignature.IMPORT_FILE_DATA.getValue());
-      importFileData.put(LiteralConstants.DATASETID, datasetId);
-      lockService.removeLockByCriteria(importFileData);
-      throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
-          EEAErrorMessage.IMPORTING_FILE_DATASET);
+    } else {
+      LOG_ERROR.error("File import failed: not multipart File");
     }
   }
